feat: Implement status options management with CRUD operations
- Added StatusOptionsController for handling HTTP requests related to status options.
- Created DTOs for status options including StatusOptionDto, CreateStatusOptionDto, and UpdateStatusOptionDto.
- Developed IStatusOptionService interface and its implementation in StatusOptionService.
- Introduced StatusOption entity to represent status options in the database.
- Implemented state machine for order and procurement status transitions to enforce valid state changes.
- Enhanced existing services (AuthService, InventoryService, NotificationService, OrderService, ProcurementService) with additional business logic and validations.
- Updated database context to include new entities and configurations.
- Added checks for existing status options to prevent duplicates during creation.
- Improved logging and error handling across services.

+++ OrdersAPI.Application/DTOs/InventoryDtos.cs
@@ -8,8 +8,8 @@ public class StoreProductDto
     public string Name { get; set; } = string.Empty;
     public string? Description { get; set; }
     public decimal PurchasePrice { get; set; }
-    public int CurrentStock { get; set; }
-    public int MinimumStock { get; set; }
+    public decimal CurrentStock { get; set; }
+    public decimal MinimumStock { get; set; }
     public string Unit { get; set; } = string.Empty;
     public bool IsLowStock { get; set; }
     public DateTime LastRestocked { get; set; }
@@ -22,8 +22,8 @@ public class CreateStoreProductDto
     public string Name { get; set; } = string.Empty;
     public string? Description { get; set; }
     public decimal PurchasePrice { get; set; }
-    public int CurrentStock { get; set; } = 0;
-    public int MinimumStock { get; set; } = 10;
+    public decimal CurrentStock { get; set; } = 0;
+    public decimal MinimumStock { get; set; } = 10;
     public string Unit { get; set; } = "pcs";
 }
 
@@ -32,8 +32,8 @@ public class UpdateStoreProductDto
     public string? Name { get; set; }
     public string? Description { get; set; }
     public decimal? PurchasePrice { get; set; }
-    public int? CurrentStock { get; set; }
-    public int? MinimumStock { get; set; }
+    public decimal? CurrentStock { get; set; }
+    public decimal? MinimumStock { get; set; }
     public string? Unit { get; set; }
 }
 
@@ -42,7 +42,7 @@ public class InventoryLogDto
     public Guid Id { get; set; }
     public Guid StoreProductId { get; set; }
     public string StoreProductName { get; set; } = string.Empty;
-    public int QuantityChange { get; set; }
+    public decimal QuantityChange { get; set; }
     public string Type { get; set; } = string.Empty;
     public string? Reason { get; set; }
     public DateTime CreatedAt { get; set; }
@@ -50,7 +50,7 @@ public class InventoryLogDto
 
 public class AdjustInventoryDto
 {
-    public int QuantityChange { get; set; }
+    public decimal QuantityChange { get; set; }
     public string Type { get; set; } = "Adjustment";
     public string? Reason { get; set; }
 }
@@ -59,7 +59,7 @@ public class ConsumptionForecastDto
 {
     public Guid StoreProductId { get; set; }
     public string StoreProductName { get; set; } = string.Empty;
-    public int CurrentStock { get; set; }
+    public decimal CurrentStock { get; set; }
     public double AverageDailyConsumption { get; set; }
     public int EstimatedDaysUntilDepletion { get; set; }
     public bool NeedsReorder { get; set; }
